+++ Controllers/Employee/EmployeeControllerDelete.cs
@@ -1,0 +1,24 @@
+using Microsoft.AspNetCore.Mvc;
+using RestAdminV2.Models;
+
+namespace RestAdmin.Controllers
+{
+    public partial class EmployeeController
+    {
+        // DELETE: api/employee/5
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteEmployee(int id)
+        {
+            var employee = await _context.Employees.FindAsync(id);
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            _context.Employees.Remove(employee);
+            await _context.SaveChangesAsync();
+
+            return NoContent();
+        }
+    }
+}
