+++ Controllers/Employee/EmployeeControllerGet.cs
@@ -1,0 +1,30 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using RestAdmin.Models;
+using RestAdminV2.Models;
+
+namespace RestAdmin.Controllers
+{
+    public partial class EmployeeController
+    {
+        // GET: api/Employee
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<Employee>>> GetEmployees()
+        {
+            return await _context.Employees.ToListAsync();
+        }
+
+        // GET: api/Employee/5
+        [HttpGet("{id}")]
+        public async Task<ActionResult<Employee>> GetEmployee(int id)
+        {
+            var employee = await _context.Employees.FindAsync(id);
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            return employee;
+        }
+    }
+}
