--- Controllers/EmployeeController.cs
@@ -1,106 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using RestAdmin.Models;
-using Microsoft.EntityFrameworkCore;
-using RestAdminV2.Models;
-
-namespace RestAdmin.Controllers;
-[ApiController]
-[Route("api/[controller]")]
-public class EmployeeController : ControllerBase
-{
-    private readonly ApplicationDbContext _context;
-
-        public EmployeeController(ApplicationDbContext context)
-        {
-            _context = context;
-        }
-    // GET: api/employee
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<Employee>>> GetEmployees()
-        {
-            var employees = await _context.Employees.ToListAsync();
-            return Ok(employees);
-        }
-    // GET: api/employee/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Employee>> GetEmployee(int id)
-        {
-            var employee = await _context.Employees.FindAsync(id);
-
-            if (employee == null)
-            {
-                return NotFound();
-            }
-
-            return Ok(employee);
-        }
-    // POST: api/employee
-        [HttpPost]
-        public async Task<ActionResult<Employee>> CreateEmployee([FromBody] Employee employee)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            _context.Employees.Add(employee);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction(nameof(GetEmployee), new { id = employee.Id }, employee);
-        }
-    // PUT: api/employee/5
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateEmployee(int id, [FromBody] Employee employee)
-        {
-            if (id != employee.Id)
-            {
-                return BadRequest("Employee ID mismatch.");
-            }
-
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            _context.Entry(employee).State = EntityState.Modified;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!EmployeeExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return NoContent();
-        }
-    // DELETE: api/employee/5
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteEmployee(int id)
-        {
-            var employee = await _context.Employees.FindAsync(id);
-            if (employee == null)
-            {
-                return NotFound();
-            }
-
-            _context.Employees.Remove(employee);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
-        }
-    private bool EmployeeExists(int id)
-        {
-            return _context.Employees.Any(e => e.Id == id);
-        }                        
-
-
-}
